changed script to generate index.md instead of index.html (WIP)
